+++ src/client/rpc_api.rs
@@ -59,8 +59,6 @@ impl RpcApi for Client {
 
         Ok(tx.compute_txid())
     }
-    /// Because there are no blocks, this function works pretty much same as
-    /// `get_transaction`.
     fn get_raw_transaction(
         &self,
         txid: &bitcoin::Txid,
@@ -68,6 +66,13 @@ impl RpcApi for Client {
     ) -> bitcoincore_rpc::Result<bitcoin::Transaction> {
         Ok(self.ledger.get_transaction(*txid)?)
     }
+    /// Verbose flag enabled `get_raw_transaction`.
+    ///
+    /// This function **is not** intended to return information about the
+    /// transaction itself. Instead, it mostly provides information about the
+    /// transaction's state in blockchain. It is recommmended to use
+    /// `get_raw_transaction` for information about transaction's inputs and
+    /// outputs.
     fn get_raw_transaction_info(
         &self,
         txid: &bitcoin::Txid,
@@ -135,9 +140,17 @@ impl RpcApi for Client {
         let tx_block_height = self
             .ledger
             .get_transaction_block_height(&tx.compute_txid())?;
+        let blockhash = match self.ledger.get_transaction_block_hash(txid) {
+            Ok(bh) => Some(bh),
+            Err(_) => None,
+        };
+        let blocktime = match self.ledger.get_block_time(tx_block_height) {
+            Ok(bt) => Some(bt as usize),
+            Err(_) => None,
+        };
         let confirmations = match self.ledger.get_mempool_transaction(*txid) {
             Some(_) => None,
-            None => Some((current_block_height - tx_block_height) as u32),
+            None => Some((current_block_height - tx_block_height + 1) as u32),
         };
 
         Ok(GetRawTransactionResult {
@@ -151,10 +164,10 @@ impl RpcApi for Client {
             locktime: 0,
             vin,
             vout,
-            blockhash: None,
+            blockhash,
             confirmations,
             time: None,
-            blocktime: None,
+            blocktime,
         })
     }
 
@@ -195,9 +208,13 @@ impl RpcApi for Client {
         let current_time = self.ledger.get_block_time(current_height)?;
         let tx_block_height = self.ledger.get_transaction_block_height(txid)?;
         let tx_block_time = self.ledger.get_block_time(tx_block_height)?;
+        let blockhash = match self.ledger.get_transaction_block_hash(txid) {
+            Ok(h) => Some(h),
+            Err(_) => None,
+        };
         let info = WalletTxInfo {
-            confirmations: (current_height - tx_block_height) as i32,
-            blockhash: None,
+            confirmations: (current_height as i64 - tx_block_height as i64 + 1) as i32,
+            blockhash,
             blockindex: None,
             blocktime: Some(current_time as u64),
             blockheight: Some(current_height),
@@ -217,8 +234,12 @@ impl RpcApi for Client {
         })
     }
 
-    /// Warning `send_to_address` won't check anything. It will only send funds
-    /// to specified address. This means: Unlimited free money.
+    /// Sends specified amount to `address` regardless of the user balance.
+    /// Meaning: Unlimited free money.
+    ///
+    /// Reason this call behaves like this is there are no wallet
+    /// implementation. This is intended way to generate inputs for other
+    /// transactions.
     fn send_to_address(
         &self,
         address: &Address<NetworkChecked>,
@@ -246,7 +267,8 @@ impl RpcApi for Client {
     }
 
     /// Creates a random secret/public key pair and generates a Bitcoin address
-    /// from witness program.
+    /// from witness program. Please note that this address is not hold in
+    /// ledger in any way.
     fn get_new_address(
         &self,
         _label: Option<&str>,
@@ -257,8 +279,9 @@ impl RpcApi for Client {
         Ok(address.as_unchecked().to_owned())
     }
 
-    /// Generates `block_num` amount of block rewards to user. Block reward is
-    /// fixed to 1 BTC, regardless of which and how many blocks are generated.
+    /// Generates `block_num` amount of block rewards to `address`. Block reward
+    /// is fixed to 1 BTC, regardless of which and how many blocks are
+    /// generated. Also mines current mempool transactions to a block.
     fn generate_to_address(
         &self,
         block_num: u64,
@@ -282,230 +305,268 @@ impl RpcApi for Client {
         Ok(vec![BlockHash::all_zeros(); block_num as usize])
     }
 
-    /// TODO: whole function
+    /// This function is intended for retrieving information about a txout's
+    /// value and confirmation. Other data may not be reliable.
+    ///
+    /// This will include mempool txouts regardless of the `include_mempool`
+    /// flag. `coinbase` will be set to false regardless if it is or not.
     fn get_tx_out(
         &self,
-        _txid: &bitcoin::Txid,
-        _vout: u32,
+        txid: &bitcoin::Txid,
+        vout: u32,
         _include_mempool: Option<bool>,
     ) -> bitcoincore_rpc::Result<Option<json::GetTxOutResult>> {
+        let current_height = self.ledger.get_block_height()?;
+        let current_block = self.ledger.get_block_with_height(current_height)?;
+        let bestblock = current_block.block_hash();
+
+        let tx = self.get_raw_transaction(txid, None)?;
+        let value = tx.output.get(vout as usize).unwrap().value;
+
+        let confirmations = self.get_transaction(txid, None)?.info.confirmations as u32;
+
         Ok(Some(GetTxOutResult {
-            bestblock: BlockHash::all_zeros(),
-            confirmations: u32::MAX,
-            value: Amount::from_sat(0x45),
+            bestblock,
+            confirmations,
+            value,
             script_pub_key: GetRawTransactionResultVoutScriptPubKey {
-                asm: "asmwhat".to_string(),
+                asm: "TODO".to_string(),
                 hex: Vec::new(),
                 req_sigs: None,
                 type_: None,
                 addresses: Vec::new(),
                 address: None,
             },
-            coinbase: true,
+            coinbase: false,
         }))
     }
-
-    // / Returns user's balance. Balance is calculated using addresses that are
-    // / generated with `get_new_address` rpc call.
-    // fn get_balance(
-    //     &self,
-    //     _minconf: Option<usize>,
-    //     _include_watchonly: Option<bool>,
-    // ) -> bitcoincore_rpc::Result<Amount> {
-    //     Ok(self.ledger.calculate_balance()?)
-    // }
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::{ledger::Ledger, Client, RpcApiWrapper};
+    use bitcoin::{Amount, Network, OutPoint, TxIn};
+    use bitcoincore_rpc::RpcApi;
+
     #[test]
-    fn raw_transaction() {
-        // let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
+    fn send_get_raw_transaction() {
+        let rpc = Client::new("send_get_raw_transaction", bitcoincore_rpc::Auth::None).unwrap();
 
-        // let credential = Ledger::generate_credential_from_witness();
-        // rpc.ledger.add_credential(credential.clone());
-        // let address = credential.address;
+        let credential = Ledger::generate_credential_from_witness();
+        let address = credential.address;
 
-        // // First, add some funds to user, for free.
-        // let txout = rpc
-        //     .ledger
-        //     .create_txout(Amount::from_sat(100_000_000), address.script_pubkey());
-        // let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
-        // let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
+        // First, create a transaction for the next txin.
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(100_000_000), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
+        let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
 
-        // // Create a new raw transactions that is valid.
-        // let txin = rpc.ledger._create_txin(txid, 0);
-        // let txout = rpc
-        //     .ledger
-        //     .create_txout(Amount::from_sat(0x45), address.script_pubkey());
-        // let inserted_tx1 = rpc.ledger.create_transaction(vec![txin], vec![txout]);
-        // rpc.send_raw_transaction(&inserted_tx1).unwrap();
+        // Create a new raw transactions that is valid.
+        let txin = TxIn {
+            previous_output: OutPoint { txid, vout: 0 },
+            witness: credential.witness.clone().unwrap(),
+            ..Default::default()
+        };
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(0x45), address.script_pubkey());
+        let inserted_tx1 = rpc.ledger.create_transaction(vec![txin], vec![txout]);
+        rpc.send_raw_transaction(&inserted_tx1).unwrap();
 
-        // let txin = rpc.ledger._create_txin(inserted_tx1.compute_txid(), 0);
-        // let txout = rpc.ledger.create_txout(
-        //     Amount::from_sat(0x45),
-        //     Ledger::generate_credential_from_witness()
-        //         .address
-        //         .script_pubkey(),
-        // );
-        // let inserted_tx2 = rpc.ledger.create_transaction(vec![txin], vec![txout]);
-        // rpc.send_raw_transaction(&inserted_tx2).unwrap();
+        let txin = TxIn {
+            previous_output: OutPoint {
+                txid: inserted_tx1.compute_txid(),
+                vout: 0,
+            },
+            witness: credential.witness.unwrap(),
+            ..Default::default()
+        };
+        let txout = rpc.ledger.create_txout(
+            Amount::from_sat(0x45),
+            Ledger::generate_credential_from_witness()
+                .address
+                .script_pubkey(),
+        );
+        let inserted_tx2 = rpc.ledger.create_transaction(vec![txin], vec![txout]);
+        rpc.send_raw_transaction(&inserted_tx2).unwrap();
 
-        // // Retrieve inserted transactions from Bitcoin.
-        // let read_tx = rpc
-        //     .get_raw_transaction(&inserted_tx1.compute_txid(), None)
-        //     .unwrap();
-        // assert_eq!(read_tx, inserted_tx1);
-        // assert_ne!(read_tx, inserted_tx2);
+        // Retrieve inserted transactions from Bitcoin.
+        let read_tx = rpc
+            .get_raw_transaction(&inserted_tx1.compute_txid(), None)
+            .unwrap();
+        assert_eq!(read_tx, inserted_tx1);
+        assert_ne!(read_tx, inserted_tx2);
 
-        // let read_tx = rpc
-        //     .get_raw_transaction(&inserted_tx2.compute_txid(), None)
-        //     .unwrap();
-        // assert_eq!(read_tx, inserted_tx2);
-        // assert_ne!(read_tx, inserted_tx1);
+        let read_tx = rpc
+            .get_raw_transaction(&inserted_tx2.compute_txid(), None)
+            .unwrap();
+        assert_eq!(read_tx, inserted_tx2);
+        assert_ne!(read_tx, inserted_tx1);
     }
 
     #[test]
-    fn transaction() {
-        // let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
+    fn get_raw_transaction_info() {
+        let rpc = Client::new("get_raw_transaction_info", bitcoincore_rpc::Auth::None).unwrap();
 
-        // let credential = Ledger::generate_credential_from_witness();
-        // rpc.ledger.add_credential(credential.clone());
-        // let address = credential.address;
+        let credential = Ledger::generate_credential_from_witness();
+        let address = credential.address;
 
-        // // First, add some funds to user, for free.
-        // let txout = rpc
-        //     .ledger
-        //     .create_txout(Amount::from_sat(100_000_000), address.script_pubkey());
-        // let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
-        // let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(0x45), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
+        let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
 
-        // // Insert raw transactions to Bitcoin.
-        // let txin = rpc.ledger._create_txin(txid, 0);
-        // let txout = rpc
-        //     .ledger
-        //     .create_txout(Amount::from_sat(0x1F), address.script_pubkey());
-        // let tx = rpc.ledger.create_transaction(vec![txin], vec![txout]);
-        // rpc.send_raw_transaction(&tx).unwrap();
+        // No blocks are mined. Little to none information is available.
+        let info = rpc.get_raw_transaction_info(&txid, None).unwrap();
+        assert_eq!(info.txid, txid);
+        assert_eq!(info.blockhash, None);
+        assert_eq!(info.confirmations, None);
 
-        // let txid = tx.compute_txid();
+        // Mining blocks should enable more transaction information.
+        rpc.ledger.mine_block().unwrap();
+        let info = rpc.get_raw_transaction_info(&txid, None).unwrap();
+        assert_eq!(info.txid, txid);
+        assert_eq!(
+            info.blockhash,
+            Some(rpc.ledger.get_transaction_block_hash(&txid).unwrap())
+        );
+        assert_eq!(info.confirmations, Some(1));
 
-        // let tx = rpc.get_transaction(&txid, None).unwrap();
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(0x1F), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
+        let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
 
-        // assert_eq!(txid, tx.info.txid);
+        // No blocks are mined. Little to none information is available.
+        let info = rpc.get_raw_transaction_info(&txid, None).unwrap();
+        assert_eq!(info.txid, txid);
+        assert_eq!(info.blockhash, None);
+        assert_eq!(info.confirmations, None);
+
+        // Mining blocks should enable more transaction information.
+        rpc.ledger.mine_block().unwrap();
+        rpc.ledger.mine_block().unwrap();
+        rpc.ledger.mine_block().unwrap();
+        let info = rpc.get_raw_transaction_info(&txid, None).unwrap();
+        assert_eq!(info.txid, txid);
+        assert_eq!(
+            info.blockhash,
+            Some(rpc.ledger.get_transaction_block_hash(&txid).unwrap())
+        );
+        assert_eq!(info.confirmations, Some(3));
+    }
+
+    #[test]
+    fn get_transaction() {
+        let rpc = Client::new("get_transaction", bitcoincore_rpc::Auth::None).unwrap();
+
+        let credential = Ledger::generate_credential_from_witness();
+        let address = credential.address;
+
+        // First, add some funds to user, for free.
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(100_000_000), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
+        let txid = rpc.ledger.add_transaction_unconditionally(tx).unwrap();
+
+        // Insert raw transactions to Bitcoin.
+        let txin = TxIn {
+            previous_output: OutPoint { txid, vout: 0 },
+            witness: credential.witness.unwrap(),
+            ..Default::default()
+        };
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(0x1F), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![txin], vec![txout]);
+        rpc.send_raw_transaction(&tx).unwrap();
+
+        let txid = tx.compute_txid();
+
+        let tx = rpc.get_transaction(&txid, None).unwrap();
+
+        assert_eq!(txid, tx.info.txid);
     }
 
     #[test]
     fn send_to_address() {
-        // let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
+        let rpc = Client::new("send_to_address", bitcoincore_rpc::Auth::None).unwrap();
 
-        // let credential = Ledger::generate_credential_from_witness();
-        // let receiver_address = credential.address;
+        let credential = Ledger::generate_credential_from_witness();
+        let receiver_address = credential.address;
 
-        // // send_to_address should send `amount` to `address`, regardless of the
-        // // user's balance.
-        // let txid = rpc
-        //     .send_to_address(
-        //         &receiver_address,
-        //         Amount::from_sat(0x45),
-        //         None,
-        //         None,
-        //         None,
-        //         None,
-        //         None,
-        //         None,
-        //     )
-        //     .unwrap();
+        // send_to_address should send `amount` to `address`, regardless of the
+        // user's balance.
+        let txid = rpc
+            .send_to_address(
+                &receiver_address,
+                Amount::from_sat(0x45),
+                None,
+                None,
+                None,
+                None,
+                None,
+                None,
+            )
+            .unwrap();
 
-        // let tx = rpc.get_raw_transaction(&txid, None).unwrap();
+        let tx = rpc.get_raw_transaction(&txid, None).unwrap();
 
-        // // Receiver should have this.
-        // assert_eq!(tx.output[0].value.to_sat(), 0x45);
-        // assert_eq!(tx.output[0].script_pubkey, receiver_address.script_pubkey());
+        // Receiver should have this.
+        assert_eq!(tx.output[0].value.to_sat(), 0x45);
+        assert_eq!(tx.output[0].script_pubkey, receiver_address.script_pubkey());
     }
 
-    // #[test]
-    // fn get_new_address() {
-    //     let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
+    #[test]
+    fn get_new_address() {
+        let rpc = Client::new("get_new_address", bitcoincore_rpc::Auth::None).unwrap();
 
-    //     let address = rpc.get_new_address(None, None).unwrap();
+        let address = rpc.get_new_address(None, None).unwrap();
 
-    //     assert!(address.is_valid_for_network(Network::Regtest));
-    //     assert!(!address.is_valid_for_network(Network::Testnet));
-    //     assert!(!address.is_valid_for_network(Network::Signet));
-    //     assert!(!address.is_valid_for_network(Network::Bitcoin));
-    //     assert_eq!(
-    //         *rpc.ledger.get_credentials()[0].address.as_unchecked(),
-    //         address
-    //     );
+        assert!(address.is_valid_for_network(Network::Regtest));
+        assert!(!address.is_valid_for_network(Network::Testnet));
+        assert!(!address.is_valid_for_network(Network::Signet));
+        assert!(!address.is_valid_for_network(Network::Bitcoin));
+    }
 
-    //     const ADDRESS_COUNT: usize = 100;
-    //     let mut prev = address;
-    //     for i in 0..ADDRESS_COUNT {
-    //         let curr = rpc.get_new_address(None, None).unwrap();
+    #[test]
+    fn generate_to_address() {
+        let rpc = Client::new("generate_to_address", bitcoincore_rpc::Auth::None).unwrap();
 
-    //         assert_ne!(prev, curr);
-    //         assert!(curr.is_valid_for_network(Network::Regtest));
-    //         assert!(!curr.is_valid_for_network(Network::Testnet));
-    //         assert!(!curr.is_valid_for_network(Network::Signet));
-    //         assert!(!curr.is_valid_for_network(Network::Bitcoin));
-    //         assert_eq!(
-    //             *rpc.ledger.get_credentials()[i + 1].address.as_unchecked(),
-    //             curr
-    //         );
+        let credential = Ledger::generate_credential_from_witness();
+        let address = credential.address;
 
-    //         prev = curr;
-    //     }
-    // }
+        // Empty wallet should reject transaction.
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(1), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
+        if let Ok(()) = rpc.ledger.check_transaction(&tx) {
+            assert!(false);
+        };
 
-    // #[test]
-    // fn generate_to_address() {
-    //     let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
+        // Generating blocks should add funds to wallet.
+        rpc.generate_to_address(101, &address).unwrap();
 
-    //     let address = rpc.get_new_address(None, None).unwrap().assume_checked();
-
-    //     // Empty wallet should reject transaction.
-    //     let txout = rpc
-    //         .ledger
-    //         .create_txout(Amount::from_sat(1), address.script_pubkey());
-    //     let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
-    //     if let Ok(()) = rpc.ledger.check_transaction(&tx) {
-    //         assert!(false);
-    //     };
-
-    //     // Generating blocks should add funds to wallet.
-    //     rpc.generate_to_address(101, &address).unwrap();
-
-    //     // Wallet has funds now. It should not be rejected.
-    //     let txin = rpc.ledger._create_txin(
-    //         rpc.ledger.get_transactions().get(0).unwrap().compute_txid(),
-    //         0,
-    //     );
-    //     let txout = rpc
-    //         .ledger
-    //         .create_txout(Amount::from_sat(1), address.script_pubkey());
-    //     let tx = rpc.ledger.create_transaction(vec![txin], vec![txout]);
-    //     if let Err(e) = rpc.ledger.check_transaction(&tx) {
-    //         assert!(false, "{:?}", e);
-    //     };
-    // }
-
-    // #[test]
-    // fn get_balance() {
-    //     let rpc = Client::new("", bitcoincore_rpc::Auth::None).unwrap();
-
-    //     let credential = Ledger::generate_credential_from_witness();
-    //     rpc.ledger.add_credential(credential.clone());
-    //     let address = credential.address;
-
-    //     assert_eq!(rpc.get_balance(None, None).unwrap(), Amount::from_sat(0));
-
-    //     let txout = rpc
-    //         .ledger
-    //         .create_txout(Amount::from_sat(0x45), address.script_pubkey());
-    //     let tx = rpc.ledger.create_transaction(vec![], vec![txout]);
-    //     rpc.ledger.add_transaction_unconditionally(tx).unwrap();
-
-    //     assert_eq!(rpc.get_balance(None, None).unwrap(), Amount::from_sat(0x45));
-    // }
+        // Wallet has funds now. It should not be rejected.
+        let txin = TxIn {
+            previous_output: OutPoint {
+                txid: rpc.ledger.get_transactions().get(0).unwrap().compute_txid(),
+                vout: 0,
+            },
+            witness: credential.witness.unwrap(),
+            ..Default::default()
+        };
+        let txout = rpc
+            .ledger
+            .create_txout(Amount::from_sat(1), address.script_pubkey());
+        let tx = rpc.ledger.create_transaction(vec![txin], vec![txout]);
+        if let Err(e) = rpc.ledger.check_transaction(&tx) {
+            assert!(false, "{:?}", e);
+        };
+    }
 }