+++ src/ledger/transactions.rs
@@ -4,7 +4,8 @@ use super::{errors::LedgerError, spending_requirements::SpendingRequirementsRetu
 use bitcoin::{
     absolute,
     consensus::{Decodable, Encodable},
-    Amount, OutPoint, ScriptBuf, Transaction, TxIn, TxOut, Txid,
+    hashes::sha256d,
+    Amount, BlockHash, OutPoint, ScriptBuf, Transaction, TxIn, TxOut, Txid,
 };
 use bitcoin_scriptexec::{ExecCtx, TxTemplate};
 use rusqlite::params;
@@ -31,9 +32,10 @@ impl Ledger {
             Err(e) => return Err(LedgerError::Transaction(e.to_string())),
         };
 
+        // Use next block height as the transaction height.
         if let Err(e) = self.database.lock().unwrap().execute(
             "INSERT INTO transactions (txid, block_height, body) VALUES (?1, ?2, ?3)",
-            params![txid.to_string(), current_block_height, body],
+            params![txid.to_string(), current_block_height + 1, body],
         ) {
             return Err(LedgerError::Transaction(format!(
                 "Couldn't add transaction with txid {} to ledger: {}",
@@ -95,6 +97,26 @@ impl Ledger {
         };
 
         Ok(sequence)
+    }
+
+    pub fn get_transaction_block_hash(&self, txid: &Txid) -> Result<BlockHash, LedgerError> {
+        let height = self.get_transaction_block_height(txid)?;
+
+        let hash = self.database.lock().unwrap().query_row(
+            "SELECT hash FROM blocks WHERE height = ?1",
+            params![height],
+            |row| Ok(row.get::<_, Vec<u8>>(0)?),
+        );
+
+        let hash = match hash {
+            Ok(h) => {
+                let hash = sha256d::Hash::consensus_decode(&mut h.as_slice()).unwrap();
+                BlockHash::from_raw_hash(hash)
+            }
+            Err(_) => return Err(LedgerError::BlockInMempool(height)),
+        };
+
+        Ok(hash)
     }
 
     pub fn get_transactions(&self) -> Vec<Transaction> {